few simple changes in Controller

+++ app/Http/Controllers/Admin/UserAccountController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Str;
 
 class UserAccountController extends Controller
@@ -14,7 +15,9 @@ class UserAccountController extends Controller
     public function index()
     {
         // Fetch all users from the database
-        $users = User::all(); 
+        $users = User::where('role', '!=', 'admin') // Exclude admin users
+                    ->get();
+        // dd($users); // Exclude admin users; 
         return inertia('Admin/listUser', ['users' => $users]);
 
     }
@@ -69,6 +72,10 @@ class UserAccountController extends Controller
     //a method to edit a user
     public function edit(User $user)
     {
+        //check if use is not an admin
+        if ($user->role === 'admin') {
+            return back()->with('error', 'You cannot edit an admin user.');
+        }
         // Return the view with the user data for editing
         return inertia('Admin/editUser', ['user' => $user]);
     }
@@ -76,6 +83,11 @@ class UserAccountController extends Controller
     //a method to update a user
     public function update(Request $request, User $user)
     {
+        Auth::check();
+         //check if use is not an admin
+        if ($user->role === 'admin') {
+            return back()->with('error', 'You cannot edit an admin user.');
+        }
         // Validate the request data
         $validatedData = $request->validate([
             'name' => 'required|string|max:255',
@@ -96,6 +108,7 @@ class UserAccountController extends Controller
     // Method to handle user account deletion
     public function destroy(Request $request)
     {
+        Auth::check();
         // Validate the request data
         $request->validate([
             'id' => 'required|exists:users,id',
@@ -104,6 +117,11 @@ class UserAccountController extends Controller
         // dd($request->all());
         // Find the user by ID and delete it
         $user = User::findOrFail($request->input('id'));
+
+         //check if use is not an admin
+        if ($user->role === 'admin') {
+            return back()->with('error', 'You cannot edit an admin user.');
+        }
         $user->delete();
 
         // Redirect back with a success message
@@ -113,6 +131,7 @@ class UserAccountController extends Controller
     //method to show all the users with doctor role
     public function selectPractitioner()
     {
+        Auth::check();
         // Fetch all users with the role of 'doctor' and is not a practitioner
         $users = User::where('role', 'doctor')
                     ->whereDoesntHave('practitioner')
